Pill extracting stateful widget from main file to it's own file
